+++ apps/nextjs-website/src/app/solutions/page.tsx
@@ -1,36 +1,14 @@
-import SolutionsTemplate from '@/components/templates/SolutionsTemplate/SolutionsTemplate';
+import SolutionListTemplate from '@/components/templates/SolutionListTemplate/SolutionListTemplate';
 import { getSolutionsList } from '@/lib/api';
 
 const Page = async () => {
   const solutionsList = await getSolutionsList();
 
   return (
-    <SolutionsTemplate
+    <SolutionListTemplate
       hero={solutionsList.hero}
       solutions={solutionsList.solutions}
-      feature={{
-        title: 'I benefici della trasformazione digitale dei servizi',
-        items: [
-          {
-            iconName: 'MarkEmailRead',
-            subtitle:
-              'I cittadini hanno un accesso più diretto e immediato alle informazioni importanti e ai servizi che possono utilizzare',
-            title: 'Comunicazione tempestiva e trasparente',
-          },
-          {
-            iconName: 'ChangeCircle',
-            subtitle:
-              'I cittadini hanno un accesso più diretto e immediato alle informazioni importanti e ai servizi che possono utilizzare',
-            title: 'Comunicazione tempestiva e trasparente',
-          },
-          {
-            iconName: 'TimerOutlined',
-            subtitle:
-              'I cittadini hanno un accesso più diretto e immediato alle informazioni importanti e ai servizi che possono utilizzare',
-            title: 'Comunicazione tempestiva e trasparente',
-          },
-        ],
-      }}
+      features={solutionsList.features}
       successStories={solutionsList.successStories}
     />
   );