+++ apps/nextjs-website/src/editorialComponents/Feature/FeatureStackItem.tsx
@@ -1,11 +1,17 @@
 import { Box, Stack, Typography } from '@mui/material';
 import Subtitle from '@/editorialComponents/Feature/Subtitle';
 import IconWrapper from '@/components/atoms/IconWrapper/IconWrapper';
+import { BlocksContent } from '@strapi/blocks-react-renderer';
+import BlocksRendererClient from '@/components/molecules/BlocksRendererClient/BlocksRendererClient';
+import Image from 'next/image';
+import { ReactNode } from 'react';
 
 export interface FeatureItem {
-  iconName: string;
+  iconName?: string;
+  iconUrl?: string;
   title: string;
-  subtitle: string;
+  subtitle?: string;
+  content?: BlocksContent;
   link?: {
     text: string;
     url: string;
@@ -26,6 +32,34 @@ export const FeatureStackItem = ({
   const isDarkMode = useDarkTheme || theme !== 'light';
   const textStyle = isDarkMode ? 'background.paper' : 'text.primary';
   const imageStyle = isDarkMode ? 'background.paper' : 'primary.main';
+
+  const renderSubtitleOrContent = (): ReactNode | null => {
+    if (item.subtitle) {
+      return item.link ? (
+        <Subtitle
+          isDarkMode={isDarkMode}
+          subtitle={item.subtitle}
+          textLink={item.link.text}
+          url={item.link.url}
+        />
+      ) : (
+        <Typography variant='body2' color='inherit'>
+          {item.subtitle}
+        </Typography>
+      );
+    }
+
+    if (item.content) {
+      return (
+        <BlocksRendererClient
+          content={item.content}
+          paragraphSx={{ color: 'white' }}
+        />
+      );
+    }
+
+    return null;
+  };
 
   return (
     <Stack
@@ -49,26 +83,18 @@ export const FeatureStackItem = ({
         }}
         color={imageStyle}
       >
-        <IconWrapper size={64} color={imageStyle} icon={item.iconName} />
+        {item.iconName && (
+          <IconWrapper size={64} color={imageStyle} icon={item.iconName} />
+        )}
+        {item.iconUrl && (
+          <Image src={item.iconUrl} alt={item.title} width={64} height={64} />
+        )}
       </Box>
       <Stack color={textStyle} spacing={1} textAlign='center'>
         <Typography color='inherit' variant='h6'>
           {item.title}
         </Typography>
-        <>
-          {!item.link ? (
-            <Typography variant='body2' color='inherit'>
-              {item.subtitle}
-            </Typography>
-          ) : (
-            <Subtitle
-              isDarkMode={isDarkMode}
-              subtitle={item.subtitle}
-              textLink={item.link.text}
-              url={item.link.url}
-            />
-          )}
-        </>
+        {renderSubtitleOrContent()}
       </Stack>
     </Stack>
   );