+++ apps/nextjs-website/src/lib/strapi/__tests__/quickStarts.test.ts
@@ -6,6 +6,7 @@ const product = {
     id: 1,
     attributes: {
       name: 'CMS APP IO',
+      shortName: 'IO',
       description: 'Test desc ',
       slug: 'app-io',
       createdAt: '2024-02-15T09:57:22.179Z',